+++ nodeconductor/cost_tracking/serializers.py
@@ -3,7 +3,7 @@ from __future__ import unicode_literals
 from rest_framework import serializers
 from django.contrib.contenttypes.models import ContentType
 
-from nodeconductor.core.serializers import GenericRelatedField, AugmentedSerializerMixin
+from nodeconductor.core.serializers import GenericRelatedField, AugmentedSerializerMixin, JSONField
 from nodeconductor.cost_tracking import models
 from nodeconductor.structure import SupportedServices
 from nodeconductor.structure import models as structure_models
@@ -99,18 +99,16 @@ class PriceListItemSerializer(serializers.HyperlinkedModelSerializer):
 
 class DefaultPriceListItemSerializer(serializers.HyperlinkedModelSerializer):
 
-    resource_content_type = serializers.SerializerMethodField()  # Deprecated, should be delete in NC-921
     resource_type = serializers.SerializerMethodField()
+    value = serializers.FloatField()
+    metadata = JSONField()
 
     class Meta:
         model = models.DefaultPriceListItem
-        fields = ('url', 'uuid', 'key', 'item_type', 'value', 'units', 'resource_content_type', 'resource_type')
+        fields = ('url', 'uuid', 'key', 'item_type', 'value', 'resource_type', 'metadata')
         extra_kwargs = {
             'url': {'lookup_field': 'uuid'},
         }
-
-    def get_resource_content_type(self, obj):
-        return '{}.{}'.format(obj.resource_content_type.app_label, obj.resource_content_type.model)
 
     def get_resource_type(self, obj):
         return SupportedServices.get_name_for_model(obj.resource_content_type.model_class())